refactor: remove serial implementations from filters and engine
