+++ filters/binarization/pipeline.go
@@ -18,6 +18,6 @@ import (
 // binarization works best on grayscaled image, it uses the GrayscalePipeline
 // as a preprocessing step.
 var BinarizationPipeline = types.FilterPipeline{
-	Preprocess:   &grayscale.GrayscalePipeline,
-	Filter:       concurrentBinarization,
+	Preprocess: grayscale.Grayscale,
+	Filter:     Binarization,
 }
