--- filters/sobel/concurrent.go
@@ -1,87 +1,0 @@
-package sobel
-
-import (
-	"image"
-	"math"
-	"sync"
-
-	"dougdomingos.com/image-filters/filters/imgutil"
-)
-
-// concurrentSobel applies the sobel filter to the entire image using multiple
-// goroutines. It makes a full copy of the image, which is shared amongst the
-// routines, which then modify their respective partitions.
-func concurrentSobel(img *image.RGBA) {
-	var (
-		bounds      = img.Bounds()
-		numWorkers  = imgutil.GetNumberOfWorkers(bounds)
-		imageStrips = imgutil.GetVerticalPartitions(bounds, numWorkers)
-		mainWg      sync.WaitGroup
-		copyWg      sync.WaitGroup
-	)
-
-	mainWg.Add(numWorkers)
-	copyWg.Add(numWorkers)
-	for strip := range imageStrips {
-		go sobelWorker(img, imageStrips[strip], &mainWg, &copyWg)
-	}
-
-	mainWg.Wait()
-}
-
-// sobelWorker processes a subregion of the image by applying the sobel filter
-// based on a global copy of the original image, which is used to compute the
-// gradients of each color channel.
-func sobelWorker(img *image.RGBA, bounds image.Rectangle, mainWg, copyWg *sync.WaitGroup) {
-	defer mainWg.Done()
-	copyImg := func() image.RGBA {
-		defer copyWg.Done()
-		return imgutil.CopyPaddedImagePartition(img, bounds, copyPadding)
-	}()
-
-	paddedMinX, paddedMaxX := copyImg.Rect.Min.X+copyPadding, copyImg.Rect.Max.X-copyPadding
-	paddedMinY, paddedMaxY := copyImg.Rect.Min.Y+copyPadding, copyImg.Rect.Max.Y-copyPadding
-
-	for y := paddedMinY; y < paddedMaxY; y++ {
-		srcRowStart := (y - copyPadding) * img.Stride
-		for x := paddedMinX; x < paddedMaxX; x++ {
-			if x == paddedMaxX-copyPadding {
-				copyWg.Wait()
-			}
-
-			offset := srcRowStart + (x-copyPadding)*4
-
-			var r8, g8, b8, a8 uint8
-			var gxR, gxG, gxB, gyR, gyG, gyB int
-
-			for ky := -1; ky <= 1; ky++ {
-				for kx := -1; kx <= 1; kx++ {
-					deltaX := x + kx
-					deltaY := y + ky
-
-					r8, g8, b8, a8 = imgutil.GetRGBA8(&copyImg, deltaX, deltaY)
-					r := int(r8)
-					g := int(g8)
-					b := int(b8)
-
-					valKernelX := gX[ky+1][kx+1]
-					valKernelY := gY[ky+1][kx+1]
-
-					gxR += int(valKernelX * r)
-					gxG += int(valKernelX * g)
-					gxB += int(valKernelX * b)
-
-					gyR += int(valKernelY * r)
-					gyG += int(valKernelY * g)
-					gyB += int(valKernelY * b)
-				}
-			}
-
-			gradR := clampColorValue(math.Sqrt(float64(gxR*gxR + gyR*gyR)))
-			gradG := clampColorValue(math.Sqrt(float64(gxG*gxG + gyG*gyG)))
-			gradB := clampColorValue(math.Sqrt(float64(gxB*gxB + gyB*gyB)))
-
-			copy(img.Pix[offset:offset+4], []uint8{gradR, gradG, gradB, a8})
-		}
-	}
-}
