+++ filters/vertical_flip/pipeline.go
@@ -7,6 +7,6 @@ package vertical_flip
 import "dougdomingos.com/image-filters/filters/types"
 
 var VerticalFlipPipeline = types.FilterPipeline{
-	Preprocess:   nil,
-	Filter:       concurrentVerticalFlip,
+	Preprocess: nil,
+	Filter:     VerticalFlip,
 }
